rewrite redstone engines, certain pipes among others; start work on #2075

+++ common/buildcraft/silicon/TileAdvancedCraftingTable.java
@@ -218,7 +218,7 @@ public class TileAdvancedCraftingTable extends TileLaserTableBase implements IIn
 
 	@Override
 	public int getProgressScaled(int i) {
-		return (int) ((getEnergy() * i) / REQUIRED_POWER);
+		return (getEnergy() * i) / REQUIRED_POWER;
 	}
 
 	@Override
